# infrastructure/swagger/openapi.yml
openapi: 3.0.3
info:
  title: Smart Banking System API
  description: |
    Complete REST API for the TT-BANK Smart Banking System.

    **Architecture:** Microservices — all requests enter through the API Gateway on port 8080.
    The gateway validates the JWT Bearer token before routing to the appropriate backend service.

    **Authentication flow:**
    1. `POST /api/v1/auth/register` — create account
    2. `POST /api/v1/auth/verify-email` — verify OTP sent by email
    3. `POST /api/v1/auth/login` — receive `accessToken` and `refreshToken`
    4. Include `Authorization: Bearer <accessToken>` on all subsequent requests

    **Currency:** Primary currency is **XAF** (Central African CFA franc). Also supports USD, EUR, GBP.

    **Base URL (local dev):** `http://localhost:8080`
    **Base URL (production):** `https://yourdomain.com`
  version: "1.0.0"
  contact:
    name: Smart Banking Dev Team
servers:
  - url: http://localhost:8080
    description: Local development (API Gateway)
  - url: https://yourdomain.com
    description: Production (replace with actual domain)

# =============================================================
# Security scheme — JWT Bearer token (HS256)
# =============================================================
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        JWT access token issued by POST /api/v1/auth/login.
        Include as: Authorization: Bearer <token>

  # =============================================================
  # Reusable schemas — built from actual Java DTOs
  # =============================================================
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
          example: "Operation completed successfully."
...
    RegisterMerchantRequest:
      type: object
      required: [businessName, businessEmail, businessPhone, businessAddress, businessCategory, walletNumber]
      properties:
        businessName:
          type: string
          minLength: 2
          maxLength: 100
          example: "Nkomo Electronics"
        businessEmail:
          type: string
          format: email
          example: "nkomo.electronics@example.com"
        businessPhone:
          type: string
          pattern: '^\\+?[0-9]{8,15}$'
          example: "+237677890123"
        businessAddress:
          type: string
          maxLength: 255
          example: "Avenue Kennedy, Yaoundé, Cameroun"
        businessCategory:
          $ref: '#/components/schemas/BusinessCategory'
        description:
          type: string
          maxLength: 500
          example: "Electronics retailer specialising in smartphones and accessories"
        walletNumber:
          type: string
          description: Existing wallet for receiving merchant payments
          example: "WLT-20240601-00001"
    MerchantPaymentRequest:
      type: object
      required: [merchantCode, customerWalletNumber, amount]
      properties:
        merchantCode:
          type: string
          description: Merchant code decoded from the QR code
          example: "MCH-20240601-XY9Z"
        customerWalletNumber:
          type: string
          description: Customer's wallet number to debit
          example: "WLT-20240601-00002"
        amount:
          type: number
          format: decimal
          minimum: 1
          example: 15000.00
        description:
          type: string
          maxLength: 255
          example: "Purchase of phone case"
    BusinessCategory:
      type: string
      enum:
        - RETAIL
        - FOOD_AND_BEVERAGE
        - HEALTHCARE
        - EDUCATION
        - TRANSPORT
        - TELECOMMUNICATIONS
        - UTILITIES
        - ENTERTAINMENT
        - SERVICES
        - OTHER
    MerchantStatus:
      type: string
      enum: [PENDING, ACTIVE, SUSPENDED, REJECTED]
    AdminStatusRequest:
      type: object
      required: [status]
      properties:
        status:
          $ref: '#/components/schemas/MerchantStatus'
        reason:
          type: string
          example: "Account verified after document review."

    # ── Savings ───────────────────────────────────────────────
...
    ErrorResponse:
      type: object
      properties:
        status:
          type: integer
          example: 400
        error:
          type: string
          example: "Bad Request"
        message:
          type: string
          example: "Validation failed: email must be a valid email address"
        timestamp:
          type: string
          format: date-time
          example: "2024-06-01T14:30:00"
        path:
          type: string
          example: "/api/v1/auth/register"
  responses:
    Unauthorized:
      description: Missing or invalid JWT Bearer token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Valid token but insufficient role (e.g. non-ADMIN accessing /api/v1/admin/**)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
security:
  - BearerAuth: []

# =============================================================
# API Paths
# =============================================================
paths:
...
  /api/v1/merchants/register:
    post:
      tags: [Merchants]
      summary: Register as a merchant
      description: >
        Registers the authenticated user as a merchant. The user must already have a wallet.
        The merchant starts in PENDING status and must be approved by an Admin.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterMerchantRequest'
      responses:
        '201':
          description: Merchant registered — awaiting admin approval
          content:
            application/json:
              schema:
                type: object
                properties:
                  merchantCode:
                    type: string
                    example: "MCH-20240601-XY9Z"
                  status:
                    $ref: '#/components/schemas/MerchantStatus'
                  message:
                    type: string
                    example: "Merchant registration submitted. Awaiting admin approval."
        '409':
          description: User is already registered as a merchant
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/merchants/me:
    get:
      tags: [Merchants]
      summary: Get the authenticated merchant's profile
      responses:
        '200':
          description: Merchant profile
          content:
            application/json:
              schema:
                type: object
                properties:
                  merchantCode:
                    type: string
                    example: "MCH-20240601-XY9Z"
                  businessName:
                    type: string
                    example: "Nkomo Electronics"
                  businessEmail:
                    type: string
                    example: "nkomo.electronics@example.com"
                  businessCategory:
                    $ref: '#/components/schemas/BusinessCategory'
                  status:
                    $ref: '#/components/schemas/MerchantStatus'
        '404':
          description: Merchant profile not found for authenticated user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/merchants/me/qr-code:
    get:
      tags: [Merchants]
      summary: Generate merchant QR code image
      description: Returns a 300×300 PNG QR code encoding the merchant's unique merchant code.
      produces:
        - image/png
      responses:
        '200':
          description: QR code PNG image
          content:
            image/png:
              schema:
                type: string
                format: binary
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/merchants/me/dashboard:
    get:
      tags: [Merchants]
      summary: Get merchant revenue dashboard data
      responses:
        '200':
          description: Dashboard statistics (total revenue, payment count, etc.)
          content:
            application/json:
              schema:
                type: object
                properties:
                  totalRevenue:
                    type: number
                    format: decimal
                    example: 1250000.00
                  totalPayments:
                    type: integer
                    example: 87
                  todayRevenue:
                    type: number
                    format: decimal
                    example: 45000.00
                  todayPayments:
                    type: integer
                    example: 3
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/merchants/me/payments:
    get:
      tags: [Merchants]
      summary: Get all payments received by the authenticated merchant
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Paginated merchant payment history
          content:
            application/json:
              schema:
                type: object
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/merchants/pay:
    post:
      tags: [Merchants]
      summary: Pay a merchant via QR code
      description: >
        Customer scans the merchant QR code and submits payment.
        Debits the customer's wallet and credits the merchant's wallet.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MerchantPaymentRequest'
      responses:
        '201':
          description: Payment successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  referenceCode:
                    type: string
                    example: "MPY-20240601-ZZZ999"
                  amount:
                    type: number
                    format: decimal
                    example: 15000.00
                  merchantName:
                    type: string
                    example: "Nkomo Electronics"
                  status:
                    type: string
                    example: "COMPLETED"
        '400':
          description: Insufficient funds or merchant not active
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/merchants/{merchantCode}:
    get:
      tags: [Merchants]
      summary: Get merchant public profile by merchant code
      parameters:
        - name: merchantCode
          in: path
          required: true
          schema:
            type: string
            example: "MCH-20240601-XY9Z"
      responses:
        '200':
          description: Merchant public profile
          content:
            application/json:
              schema:
                type: object
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/merchants/admin/all:
    get:
      tags: [Merchants, Admin]
      summary: Get all merchants (Admin only)
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/MerchantStatus'
      responses:
        '200':
          description: Paginated list of all merchants
          content:
            application/json:
              schema:
                type: object
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/merchants/admin/{merchantId}/status:
    put:
      tags: [Merchants, Admin]
      summary: Update merchant status (Admin only)
      description: Approve (ACTIVE), suspend (SUSPENDED), or reject (REJECTED) a merchant.
      parameters:
        - name: merchantId
          in: path
          required: true
          schema:
            type: integer
            format: int64
            example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminStatusRequest'
      responses:
        '200':
          description: Merchant status updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ============================================================
  # SAVINGS SERVICE — /api/v1/savings/**
  # ============================================================
...